api/requirements.txt
black==24.8.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlalchemy==2.0.40
tomli==2.2.1
typing-extensions==4.13.1
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
